+++ cli/commands/up.sh
@@ -87,7 +87,7 @@ orcan_cmd_up() {
     orcan_require_docker
     orcan_require_generated
     orcan_load_env
-    orcan_maybe_hint_update
+    orcan_maybe_hint_upgrade
 
     local git_overlay=""